Assignment2/requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
boto3==1.26.8
botocore==1.29.8
cached-property==1.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ec2-metadata==2.10.0
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
typing-extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
